Move k-counter
