+++ panel/src/ui/components/Counter.vue
